:hammer: Refactor: community 리팩토링
