+++ src/components/Community/ArticleList/ArticleList.scss
@@ -1,3 +1,11 @@
+.article-section {
+  display: flex;
+  flex-direction: column;
+  
+  width: 100%;
+  min-height: 700px;
+}
+
 .Article-list {
   width: 100%;
