--- src/components/Community/ArticleSection/ArticleSection.scss
@@ -1,7 +1,0 @@
-.article-section {
-  display: flex;
-  flex-direction: column;
-  
-  width: 100%;
-  min-height: 700px;
-}
